feat: extract common behavior from ProductCreate and Edit to ProductCreateAndEditForm
